website: First draft of "Refactoring" documentation, about "moved" blocks
This is an early iteration of documentation about the refactoring features
we're intending to introduce in a forthcoming release.

I've named the documentation section "refactoring" because in previous
discussions with users that seems to be the term they use to describe the
underlying need. "moved" blocks are our first language feature intended to
meet that need, although it probably won't be the last as we consider
other requirements in later releases. My intent here is that once we've
published this it should eventually end up being the first result for a
web search for the topic of Terraform refactoring.